[FIX] hr_timesheet_sheet: keep timesheets linked to a task
When deleting an hr_timesheet_sheet.sheet, the linked analytic lines are
dropped too.
This was introduced in 8.0 at 5f45e7ca

The initial bug, reported at #5326, was targeting the analytic lines created
independantly, directly from the timesheet encoding.

In version 8, the timesheets were encoded via project.task.work and the unlink
on the hr_timesheet_sheet.sheet had no effect on the works.

In version 10, this model no longer exists and it needs to be filtered to avoid
removing independant timesheets.

Fixes #15741