[FIX] sale: translate default name at creation
When a sale order is created from the interface, the name field is read-only so
not given in the create call.

However, if the create values are filled with the default value, the string
'New' is translated and the condition is not met in the create method.

This commit makes the create consistent with the default `lambda self: _('New')`

Fixes #15839