[FIX] account: usability of entry sequences.
This revert commits 51882ea70cb599aec8b424b0bde4edc425299c36 and d94c449547a2a7a6e270e4602320f588c5b6f896 that were introduced to solve OPW-693955.

The usability before those fixes was good, but the bug was appearing when we first create the journal without the boolean 'dedicated refund sequence' set, then we try to modify that va$

This time, the fix consist to create automatically the sequence on the write of the boolean if it's not existing yet, in the same fashion than on the create().